Big Update -> DB Class, Storage Classes, Stored Procedures
Re-added in the sitemap and footer (I think they got lost in one of the
merges). Added in some styles to main to support the footer. Created a
new PlantDB.php class -- Acts as a db class for all of the stored
procedures living on phpmyadmin. Added in storage classes for the db
class to use, essentially it creates a list of objects that it jsonifies
and kicks back out. Added in functionality to autocompletes and the
procedures that back them along with a semi functional example. Added in
support for the interactive map. Added in a tester page that shows how
to reference the db class. Added in content to the index.